requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
astropy==4.0.1.post1
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.4
bleach==3.1.5
blinker==1.4
cachetools==4.1.1
certifi==2016.2.28
cffi==1.11.5
chardet==3.0.4
click==6.7
confuse==1.3.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.11.1
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libtiff==0.4.2
llvmlite==0.33.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.3.0
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.1.1
numba==0.50.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.40
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.1
pdf2image==1.13.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.5
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.11.0
SQLAlchemy==1.2.7
tangled-up-in-unicode==0.0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tokenizers==0.8.1rc1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.2
urllib3==1.25.10
visions==0.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.1
zipp==3.1.0
